refactored shadaer code

+++ khulagl/VertexArray.cpp
@@ -33,7 +33,7 @@ void VertexArray::addBuffer(const VertexBuffer &vb, const VertexBufferLayout &la
     unsigned int offset = 0;
     for (int i = 0; i < elements.size(); i++) {
         glEnableVertexAttribArray(i);
-        glVertexAttribPointer(i, elements[i].count, elements[i].type, elements[i].normalized, layout.getStride(), (const void*) offset);
+        glVertexAttribPointer(i, elements[i].count, elements[i].type, elements[i].normalized, layout.getStride(), (const void*) (size_t) offset);
         
         offset += elements[i].size();
     }
